add pretokenized pile to normal script

+++ pkg/infini_gram/indexing_v6.py
@@ -17,233 +17,8 @@ import struct
 import zstandard as zstd
 transformers.utils.logging.set_verbosity(40) # suppress warnings
 
-# TODO: change to .dataset to import from local utility function (note: this added code if from the polypythia seed huggingface repo)
 import struct
-from functools import lru_cache
-from itertools import accumulate
-from pathlib import Path
-from torch.utils.data import Dataset
-
-dtypes = {1: np.uint8, 2: np.int8, 3: np.int16, 4: np.int32, 5: np.int64, 6: np.float32, 7: np.float64, 8: np.uint16}
-
-
-def index_file_path(prefix_path):
-    return prefix_path + ".idx"
-
-
-def data_file_path(prefix_path):
-    return prefix_path + ".bin"
-
-
-def _warmup_mmap_file(path):
-    with open(path, "rb") as stream:
-        while stream.read(100 * 1024 * 1024):
-            pass
-
-
-class MMapIndexedDataset(Dataset):
-    class Index:
-        _HDR_MAGIC = b"MMIDIDX\x00\x00"
-
-        @classmethod
-        def writer(cls, path, dtype):
-            class _Writer:
-                def __enter__(self):
-                    self._file = open(path, "wb")  # noqa: SIM115
-
-                    # Write Magic string so we can check the file format then opening it again.
-                    self._file.write(cls._HDR_MAGIC)
-                    # Write version number
-                    # Little endian unsigned 64 Bit integer
-                    self._file.write(struct.pack("<Q", 1))
-                    # Little endian unsigned 8 Bit integer
-                    self._file.write(struct.pack("<B", code(dtype)))
-
-                    return self
-
-                @staticmethod
-                def _get_pointers(sizes):
-                    pointers = np.zeros(len(sizes), dtype=np.int64)
-                    sizes = np.array(sizes, dtype=np.int64)
-
-                    np.cumsum(sizes[:-1], out=pointers[1:])
-                    pointers = pointers * dtype().itemsize
-                    return pointers
-
-                def write(self, sizes, doc_idx):
-                    pointers = self._get_pointers(sizes)
-
-                    # Little endian unsigned 64 Bit integer
-                    self._file.write(struct.pack("<Q", len(sizes)))
-                    # Little endian unsigned 64 Bit integer
-                    self._file.write(struct.pack("<Q", len(doc_idx)))
-
-                    sizes = np.array(sizes, dtype=np.int32)
-                    self._file.write(sizes.tobytes(order="C"))
-                    del sizes
-
-                    pointers = np.array(pointers, dtype=np.int64)
-                    self._file.write(pointers.tobytes(order="C"))
-                    del pointers
-
-                    doc_idx = np.array(doc_idx, dtype=np.int64)
-                    self._file.write(doc_idx.tobytes(order="C"))
-
-                def __exit__(self, exc_type, exc_val, exc_tb):
-                    self._file.close()
-
-            return _Writer()
-
-        def __init__(self, path, skip_warmup=False):
-            with open(path, "rb") as stream:
-                magic_test = stream.read(9)
-                assert magic_test == self._HDR_MAGIC, (
-                    "Index file doesn't match expected format. " "Make sure that --dataset-impl is configured properly."
-                )
-                # Little endian unsigned 64 Bit integer
-                version = struct.unpack("<Q", stream.read(8))
-                assert version == (1,)
-
-                # Little endian unsigned 8 Bit integer
-                (dtype_code,) = struct.unpack("<B", stream.read(1))
-                self._dtype = dtypes[dtype_code]
-                self._dtype_size = self._dtype().itemsize
-
-                self._len = struct.unpack("<Q", stream.read(8))[0]
-                self._doc_count = struct.unpack("<Q", stream.read(8))[0]
-                offset = stream.tell()
-
-            if not skip_warmup:
-                print("    warming up index mmap file...")
-                _warmup_mmap_file(path)
-
-            self._bin_buffer_mmap = np.memmap(path, mode="r", order="C")
-            self._bin_buffer = memoryview(self._bin_buffer_mmap)
-            print("    reading sizes...")
-            self._sizes = np.frombuffer(self._bin_buffer, dtype=np.int32, count=self._len, offset=offset)
-            print("    reading pointers...")
-            self._pointers = np.frombuffer(
-                self._bin_buffer, dtype=np.int64, count=self._len, offset=offset + self._sizes.nbytes
-            )
-            print("    reading document index...")
-            self._doc_idx = np.frombuffer(
-                self._bin_buffer,
-                dtype=np.int64,
-                count=self._doc_count,
-                offset=offset + self._sizes.nbytes + self._pointers.nbytes,
-            )
-
-        def __del__(self):
-            self._bin_buffer_mmap._mmap.close()
-            del self._bin_buffer_mmap
-
-        @property
-        def dtype(self):
-            return self._dtype
-
-        @property
-        def sizes(self):
-            return self._sizes
-
-        @property
-        def doc_idx(self):
-            return self._doc_idx
-
-        @lru_cache(maxsize=8)  # noqa: B019
-        def __getitem__(self, i):
-            return self._pointers[i], self._sizes[i]
-
-        def __len__(self):
-            return self._len
-
-    def __init__(self, path, skip_warmup=False):
-        super().__init__()
-
-        self._path = None
-        self._index = None
-        self._bin_buffer = None
-
-        self._do_init(path, skip_warmup)
-
-    def __getstate__(self):
-        return self._path
-
-    def __setstate__(self, state):
-        self._do_init(state)
-
-    def _do_init(self, path, skip_warmup):
-        self._path = path
-        self._index = self.Index(index_file_path(self._path), skip_warmup)
-
-        if not skip_warmup:
-            print("    warming up data mmap file...")
-            _warmup_mmap_file(data_file_path(self._path))
-        print("    creating numpy buffer of mmap...")
-        self._bin_buffer_mmap = np.memmap(data_file_path(self._path), mode="r", order="C")
-        print("    creating memory view of numpy buffer...")
-        self._bin_buffer = memoryview(self._bin_buffer_mmap)
-
-    def __del__(self):
-        self._bin_buffer_mmap._mmap.close()
-        del self._bin_buffer_mmap
-        del self._index
-
-    def __len__(self):
-        return len(self._index)
-
-    # @lru_cache(maxsize=8)
-    def __getitem__(self, idx):
-        if isinstance(idx, int):
-            ptr, size = self._index[idx]
-            np_array = np.frombuffer(self._bin_buffer, dtype=self._index.dtype, count=size, offset=ptr)
-            return np_array
-        elif isinstance(idx, slice):
-            start, stop, step = idx.indices(len(self))
-            if step != 1:
-                raise ValueError("Slices into indexed_dataset must be contiguous")
-            ptr = self._index._pointers[start]
-            sizes = self._index._sizes[idx]
-            offsets = list(accumulate(sizes))
-            total_size = sum(sizes)
-            np_array = np.frombuffer(self._bin_buffer, dtype=self._index.dtype, count=total_size, offset=ptr)
-            sents = np.split(np_array, offsets[:-1])
-            return sents
-
-    def get(self, idx, offset=0, length=None):
-        """Retrieves a single item from the dataset with the option to only
-        return a portion of the item.
-
-        get(idx) is the same as [idx] but get() does not support slicing.
-        """
-        ptr, size = self._index[idx]
-        if length is None:
-            length = size - offset
-        ptr += offset * np.dtype(self._index.dtype).itemsize
-        np_array = np.frombuffer(self._bin_buffer, dtype=self._index.dtype, count=length, offset=ptr)
-        return np_array
-
-    @property
-    def sizes(self):
-        return self._index.sizes
-
-    @property
-    def doc_idx(self):
-        return self._index.doc_idx
-
-    def get_doc_idx(self):
-        return self._index._doc_idx
-
-    def set_doc_idx(self, doc_idx_):
-        self._index._doc_idx = doc_idx_
-
-    @property
-    def supports_prefetch(self):
-        return False
-
-    @staticmethod
-    def exists(path):
-        return os.path.exists(index_file_path(path)) and os.path.exists(data_file_path(path))
-
+from .dataset import MMapIndexedDataset
 
 tokenizer = None
 
@@ -584,6 +359,8 @@ def build_sa(args):
 
     ds_path = os.path.join(args.save_dir, f'tokenized')
     sa_path = ds_path.replace('tokenized', 'table')
+    merged_dir = os.path.join(args.temp_dir, f'merged')
+    parts_dir = os.path.join(args.temp_dir, f'parts')
     if os.path.exists(sa_path):
         print(f'Step 2 (build_sa): Skipped. File already exists.', flush=True)
         return
@@ -594,52 +371,52 @@ def build_sa(args):
     start_time_all = time.time()
 
     # -------- Step 2.1 (make-part) -------- #
-
-    print(f'Step 2.1 (make-part): Starting ...', flush=True)
-    start_time = time.time()
-
     ds_size = os.path.getsize(ds_path)
     ratio = int(np.ceil(np.log2(ds_size) / 8))
-    mem_bytes = args.mem * 1024**3
-    num_job_batches = 1
-    while num_job_batches * (mem_bytes // (12 if args.token_width == 1 else 8)) < ds_size:
-        num_job_batches *= 2
-    parallel_jobs = args.cpus
-    total_jobs = num_job_batches * parallel_jobs
-    print(f'Using {num_job_batches} batches of {parallel_jobs} jobs each, for a total of {total_jobs} jobs.', flush=True)
+    print(f'Step 2.1 (make-part): Starting ...', flush=True)
+    if not os.path.exists(merged_dir):
+        start_time = time.time()
 
-    S = ds_size // total_jobs
-    # Make sure that parts contain whole tokens
-    if S % args.token_width != 0:
-        S += args.token_width - S % args.token_width
+        ds_size = os.path.getsize(ds_path)
+        ratio = int(np.ceil(np.log2(ds_size) / 8))
+        mem_bytes = args.mem * 1024**3
+        num_job_batches = 1
+        while num_job_batches * (mem_bytes // (12 if args.token_width == 1 else 8)) < ds_size:
+            num_job_batches *= 2
+        parallel_jobs = args.cpus
+        total_jobs = num_job_batches * parallel_jobs
+        print(f'Using {num_job_batches} batches of {parallel_jobs} jobs each, for a total of {total_jobs} jobs.', flush=True)
 
-    parts_dir = os.path.join(args.temp_dir, f'parts')
-    shutil.rmtree(parts_dir, ignore_errors=True)
-    os.makedirs(parts_dir)
+        S = ds_size // total_jobs
+        # Make sure that parts contain whole tokens
+        if S % args.token_width != 0:
+            S += args.token_width - S % args.token_width
 
-    for batch_start in tqdm(list(range(0, total_jobs, parallel_jobs))):
-        batch_end = min(batch_start+parallel_jobs, total_jobs)
-        batch_ranges = []
-        for i in range(batch_start, batch_end):
-            s, e = i*S, min((i+1)*S+args.hack, ds_size)
-            batch_ranges.append((s, e))
-        pipes = []
-        for (s, e) in batch_ranges:
-            pipes.append(os.popen(f'./rust_indexing make-part --data-file {ds_path} --parts-dir {parts_dir} --start-byte {s} --end-byte {e} --ratio {ratio} --token-width {args.token_width}'))
-        [pipe.read() for pipe in pipes]
-        if any([pipe.close() is not None for pipe in pipes]):
-            print('Step 2.1 (make-part): Something went wrong', flush=True)
-            exit(1)
+        shutil.rmtree(parts_dir, ignore_errors=True)
+        os.makedirs(parts_dir)
 
-    end_time = time.time()
-    print(f'Step 2.1 (make-part): Done. Took {end_time-start_time:.2f} seconds', flush=True)
+        for batch_start in tqdm(list(range(0, total_jobs, parallel_jobs))):
+            batch_end = min(batch_start+parallel_jobs, total_jobs)
+            batch_ranges = []
+            for i in range(batch_start, batch_end):
+                s, e = i*S, min((i+1)*S+args.hack, ds_size)
+                batch_ranges.append((s, e))
+            pipes = []
+            for (s, e) in batch_ranges:
+                pipes.append(os.popen(f'./rust_indexing make-part --data-file {ds_path} --parts-dir {parts_dir} --start-byte {s} --end-byte {e} --ratio {ratio} --token-width {args.token_width}'))
+            [pipe.read() for pipe in pipes]
+            if any([pipe.close() is not None for pipe in pipes]):
+                print('Step 2.1 (make-part): Something went wrong', flush=True)
+                exit(1)
+
+        end_time = time.time()
+        print(f'Step 2.1 (make-part): Done. Took {end_time-start_time:.2f} seconds', flush=True)
 
     # -------- Step 2.2 (merge) -------- #
 
     print(f'Step 2.2 (merge): Starting ...', flush=True)
     start_time = time.time()
 
-    merged_dir = os.path.join(args.temp_dir, f'merged')
     shutil.rmtree(merged_dir, ignore_errors=True)
     os.makedirs(merged_dir)
